Update Guard naming convention
