+++ angular/src/services/lock.guard.ts
@@ -4,11 +4,12 @@ import { CanActivate, Router } from "@angular/router";
 import { StateService } from "jslib-common/abstractions/state.service";
 import { VaultTimeoutService } from "jslib-common/abstractions/vaultTimeout.service";
 
-import { BaseGuardService } from "./base-guard.service";
+import { BaseGuard } from "./base.guard";
 
 @Injectable()
-export class UnauthGuardService extends BaseGuardService implements CanActivate {
+export class LockGuard extends BaseGuard implements CanActivate {
   protected homepage = "vault";
+  protected loginpage = "login";
   constructor(
     router: Router,
     private vaultTimeoutService: VaultTimeoutService,
@@ -18,15 +19,14 @@ export class UnauthGuardService extends BaseGuardService implements CanActivate 
   }
 
   async canActivate() {
-    const isAuthed = await this.stateService.getIsAuthenticated();
-    if (isAuthed) {
-      const locked = await this.vaultTimeoutService.isLocked();
-      if (locked) {
-        return this.redirect("lock");
-      } else {
-        return this.redirect(this.homepage);
-      }
+    if (await this.vaultTimeoutService.isLocked()) {
+      return true;
     }
-    return true;
+
+    const redirectUrl = (await this.stateService.getIsAuthenticated())
+      ? [this.homepage]
+      : [this.loginpage];
+
+    return this.redirect(redirectUrl);
   }
 }
